refactor(bridge): split code for h3_to_h2 and h3_to_h1
